LayoutTests/http/tests/websocket/tests/hybi/bad-sub-protocol-control-chars-expected.txt

CONSOLE MESSAGE: line 1: Wrong protocol for WebSocket '\u0000'
CONSOLE MESSAGE: line 1: Wrong protocol for WebSocket '\u0009'
CONSOLE MESSAGE: line 1: Wrong protocol for WebSocket '\u001B'
Test WebSocket bad sub-protocol names by control characters.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS new WebSocket("ws://127.0.0.1:8880/simple", "\0") threw exception SyntaxError (DOM Exception 12): The string did not match the expected pattern..
PASS new WebSocket("ws://127.0.0.1:8880/simple", "	") threw exception SyntaxError (DOM Exception 12): The string did not match the expected pattern..
PASS new WebSocket("ws://127.0.0.1:8880/simple", "
") threw exception SyntaxError: Unexpected EOF.
PASS new WebSocket("ws://127.0.0.1:8880/simple", "
") threw exception SyntaxError: Unexpected EOF.
PASS new WebSocket("ws://127.0.0.1:8880/simple", "") threw exception SyntaxError (DOM Exception 12): The string did not match the expected pattern..
PASS successfullyParsed is true

TEST COMPLETE
